# api/openapi/tars-mvp.yaml
openapi: 3.1.0
info:
  title: TARS MVP API
  version: 0.1.0
  description: |
    Draft API definition aligned with tars_technical_design.md v1.4.
    Tag names describe API contract groupings, not Web Console page ownership.
    In particular, the `ops` tag means authenticated control-plane APIs and is
    not equivalent to the `/ops` UI page; some `ops`-tagged APIs are primarily
    surfaced from other daily object pages such as `/identity/*`, `/providers`,
    and `/connectors`.
servers:
  - url: https://tars.example.com
...
paths:
...
  /api/v1/users:
    get:
      tags: [ops]
      summary: List users
      operationId: listUsers
      security:
        - OpsBearer: []
      parameters:
        - name: q
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
      responses:
        "200":
          description: User list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [ops]
      summary: Create user
      operationId: createUser
      security:
        - OpsBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpsertRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/ValidationFailed"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/users/{user_id}:
    get:
      tags: [ops]
      summary: Get user
      operationId: getUser
      security:
        - OpsBearer: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [ops]
      summary: Update user
      operationId: updateUser
      security:
        - OpsBearer: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpsertRequest"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/ValidationFailed"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/v1/groups:
    get:
      tags: [ops]
      summary: List groups
      operationId: listGroups
      security:
        - OpsBearer: []
      parameters:
        - name: q
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
      responses:
        "200":
          description: Group list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [ops]
      summary: Create group
      operationId: createGroup
      security:
        - OpsBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GroupUpsertRequest"
      responses:
        "201":
          description: Group created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
        "400":
          $ref: "#/components/responses/ValidationFailed"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/groups/{group_id}:
    get:
      tags: [ops]
      summary: Get group
      operationId: getGroup
      security:
        - OpsBearer: []
      parameters:
        - name: group_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Group detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [ops]
      summary: Update group
      operationId: updateGroup
      security:
        - OpsBearer: []
      parameters:
        - name: group_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GroupUpsertRequest"
      responses:
        "200":
          description: Updated group
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
        "400":
          $ref: "#/components/responses/ValidationFailed"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/v1/roles:
    get:
      tags: [ops]
      summary: List roles
      operationId: listRoles
      security:
        - OpsBearer: []
      responses:
        "200":
          description: Role list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoleListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [ops]
      summary: Create role
      operationId: createRole
      security:
        - OpsBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleUpsertRequest"
      responses:
        "201":
          description: Role created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Role"
        "400":
          $ref: "#/components/responses/ValidationFailed"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/roles/{role_id}:
    get:
      tags: [ops]
      summary: Get role
      operationId: getRole
      security:
        - OpsBearer: []
      parameters:
        - name: role_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Role detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Role"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [ops]
      summary: Update role
      operationId: updateRole
      security:
        - OpsBearer: []
      parameters:
        - name: role_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleUpsertRequest"
      responses:
        "200":
          description: Updated role
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Role"
        "400":
          $ref: "#/components/responses/ValidationFailed"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/roles/{role_id}/bindings:
    get:
      tags: [ops]
      summary: Get role bindings (users and groups bound to a role)
      operationId: getRoleBindings
      security:
        - OpsBearer: []
      parameters:
        - name: role_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Role bindings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoleBindingsResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      tags: [ops]
      summary: Bind role to users and groups
      operationId: bindRole
      security:
        - OpsBearer: []
      parameters:
        - name: role_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleBindingRequest"
      responses:
        "200":
          description: Bound role
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Role"
        "400":
          $ref: "#/components/responses/ValidationFailed"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/people:
    get:
      tags: [ops]
      summary: List people
      operationId: listPeople
      security:
        - OpsBearer: []
      parameters:
        - name: q
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
      responses:
        "200":
          description: Person list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PersonListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [ops]
      summary: Create person
      operationId: createPerson
      security:
        - OpsBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PersonUpsertRequest"
      responses:
        "201":
          description: Person created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Person"
        "400":
          $ref: "#/components/responses/ValidationFailed"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/people/{person_id}:
    get:
      tags: [ops]
      summary: Get person
      operationId: getPerson
      security:
        - OpsBearer: []
      parameters:
        - name: person_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Person detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Person"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [ops]
      summary: Update person
      operationId: updatePerson
      security:
        - OpsBearer: []
      parameters:
        - name: person_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PersonUpsertRequest"
      responses:
        "200":
          description: Updated person
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Person"
        "400":
          $ref: "#/components/responses/ValidationFailed"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    OpsBearer:
      type: http
      scheme: bearer
      bearerFormat: opaque
  responses:
    Unauthorized:
      description: Authentication required or token invalid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    ValidationFailed:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
  schemas:
...
    UserListResponse:
      allOf:
        - $ref: "#/components/schemas/ListPage"
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/User"
...
    GroupListResponse:
      allOf:
        - $ref: "#/components/schemas/ListPage"
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/Group"
...
    RoleListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Role"
...
    UserUpsertRequest:
      type: object
      required: [user]
      properties:
        user:
          $ref: "#/components/schemas/User"
        operator_reason:
          type: string
          description: Optional on IAM day-to-day user pages; backend may infer or default the audit reason.
    GroupUpsertRequest:
      type: object
      required: [group]
      properties:
        group:
          $ref: "#/components/schemas/Group"
        operator_reason:
          type: string
          description: Optional on IAM day-to-day group pages; backend may infer or default the audit reason.
...